Add research bar wich redirect to the found page
