+++ src/Controller/EventController.php
@@ -5,7 +5,6 @@ namespace App\Controller;
 use App\Entity\Event;
 use App\Repository\EventRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -30,37 +29,5 @@ class EventController extends AbstractController
         return $this->render('event/view.html.twig', [
             'event' => $oneEvent,
         ]);
-    }
-
-    /**
-     * @param Request $request
-     * @return JsonResponse
-     */
-    public function searchBarAjax(Request $request) : JsonResponse
-    {
-        $research = $request->query->get('data');
-        $doctrine = $this->getDoctrine();
-
-        /** @var EventRepository $eventRepository */
-        $eventRepository = $doctrine->getRepository(Event::class);
-
-        /** @var Event $events */
-        $events = $eventRepository->findAll();
-
-        $titles = [];
-
-        /**
-         * @var int $key
-         * @var Event $event
-         */
-        foreach ($events as $key => $event) {
-            array_push($titles, $event->getTitle());
-        }
-
-        return new JsonResponse(
-            array(
-                $titles,
-            )
-        );
     }
 }
